All cmd -e/--n-emissions now take size_t argument

+++ src/2d/barrel/matrix_cmd.cpp
@@ -149,7 +149,7 @@ static void run(cmdline::parser& cl, ModelArgs... args) {
   auto& n_pixels = cl.get<int>("n-pixels");
   auto& m_pixel = cl.get<int>("m-pixel");
   auto& s_pixel = cl.get<double>("s-pixel");
-  auto& n_emissions = cl.get<int>("n-emissions");
+  auto& n_emissions = cl.get<size_t>("n-emissions");
   auto& tof_step = cl.get<double>("tof-step");
   auto verbose = cl.count("verbose");
