+++ src/2d/strip/phantom_cmd.cpp
@@ -99,7 +99,7 @@ int main(int argc, char* argv[]) {
   }
 #endif
 
-  auto n_emissions = cl.get<int>("n-emissions");
+  auto n_emissions = cl.get<size_t>("n-emissions");
   auto verbose = cl.count("verbose");
 
   Scanner scanner(PET2D_STRIP_SCANNER_CL(cl));
